Fix code that was ensuring unnamed module can be read by compiled modules + refactoring
