+++ build.gradle
@@ -76,19 +76,20 @@ def launcher8 = javaToolchains.launcherFor { languageVersion.set(JavaLanguageVer
 def javaHome8 = launcher8.map { it.metadata.installationPath.asFile }
 def jvm8 = Jvm.forHome(javaHome8.get())
 
-dependencies {
-    implementation files(jvm8.toolsJar)
-
-    api project(":dd-javac-plugin-client")
-    api "org.burningwave:core:12.62.7"
-
-    testImplementation "org.junit.jupiter:junit-jupiter-api:5.9.2"
-    testImplementation 'org.junit.jupiter:junit-jupiter-params:5.9.2'
-    testImplementation 'commons-io:commons-io:2.11.0'
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:5.9.2"
-}
-
 apply from: "$rootDir/gradle/java8-compile.gradle"
 apply from: "$rootDir/gradle/multi-release.gradle"
 apply from: "$rootDir/gradle/publishing.gradle"
 apply from: "$rootDir/gradle/test.gradle"
+
+dependencies {
+    implementation files(jvm8.toolsJar)
+
+    api project(":dd-javac-plugin-client")
+
+    java9Implementation "org.burningwave:core:12.65.1"
+
+    testImplementation "org.junit.jupiter:junit-jupiter-api:5.9.2"
+    testImplementation "org.junit.jupiter:junit-jupiter-params:5.9.2"
+    testImplementation "commons-io:commons-io:2.11.0"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:5.9.2"
+}
