+++ src/main/java/datadog/compiler/CompilerModuleOpener.java
@@ -1,0 +1,9 @@
+package datadog.compiler;
+
+public class CompilerModuleOpener {
+
+    public static void setup() {
+        // this is a stub used only for JDK 8. It does nothing. See corresponding class in JDK 9+ sources.
+    }
+
+}
