--- src/main/java/datadog/compiler/UnnamedModuleOpener.java
@@ -1,10 +1,0 @@
-package datadog.compiler;
-
-import com.sun.tools.javac.util.Context;
-
-public class UnnamedModuleOpener {
-
-    public static void open(Context context) {
-        // no op
-    }
-}
